[minor/compile] kill warning in a portable way. The proper fix (use %td in the format instead of %d / %ld) is only available from glibc 2.1 on, so can't be used portably
git-svn-id: https://subversion.renater.fr/wimsdev/trunk@3814 b48fb02d-6135-0410-b1dd-deb767c81da9

+++ wims/src/Misc/whirlgif/gifdecod.c
@@ -120,8 +120,8 @@ void GifDecode(FILE *fp, UBYTE *pix, GifImageHdr gifimage)
   }
 
 skipRest: 
-  if (debugFlag) fprintf(stderr, "Ending GifDecode, written: %ld=%d\n",
-	  interlaced && (pix-picture == 0) ? imgsize : pix - picture, imgsize);
+  if (debugFlag) fprintf(stderr, "Ending GifDecode, written: %d=%d\n",
+	  interlaced && (pix-picture == 0) ? imgsize : (int)(pix - picture), imgsize);
   return ;
 }
